requirements.txt
flask
flask-cors
flask_sqlalchemy
openai
python-dotenv
firebase-admin